buildconfig: disable tests for undecoded keys for now
This commit partially reverts PT#549 to unblock filesystem
customizations in bib.

This is a short term fix and we should revert and do something
smarter like https://github.com/osbuild/images/pull/951 or see
if we can do better in the toml unmarshaling. But to unblock
toml customizations this is a (IMHO) reasonable first step.

Closes: https://github.com/osbuild/bootc-image-builder/issues/655