container: run `dnf check` in container for dnf4/dnf5 compat
This is an alternative idea to PR#651 - we run `dnf` inside the
container to init the dnf system. However with `dnf5` this no
longer works and it just throws an error.

So instead we run `dnf check --duplicates` which should work on
both dnf{4,5} and is reasonable fast. Just running `dnf check`
works as well but with dnf5 takes around 20s on my system and
`dnf5 check --duplicates` only 0.7s.

Co-authored-by: Hristo Marinov <hricky@mail.bg>