test: fix the testcase strings to not include `osinfo_template`
This commit fixes the testcase string representation to not include
the `osinfo_template` string and adds a test case to not accidentally
regress here. Currently our test case strings look like:
```
<Dir bootc-image-builder>
  <Dir test>
    <Module test_build.py>
      <Function test_container_builds>
      <Function test_image_is_generated[quay.io/centos-bootc/centos-bootc:stream9,qcow2+raw+vmdk+vhd+gce,CentOS Stream 9 ({arch})]>
      <Function test_image_boots[quay.io/centos-bootc/centos-bootc:stream9,raw,CentOS Stream 9 ({arch})]>
...
```
which is an accident. A testcase should ideally only contain what

Instead of putting the osinfo_template into the testcase itself,
have a function that generates it. It has downsides (now the
osinfo_template is more disconnected from the testcase) so we could
move it also back into testcase or we could exclude osinfo_template
from the string generation. Ideas welcome here, my current approach
feels okay but not like it's perfect yet.