bib: extract common `platformFor()` helper and add tests
This commit extracts a common `platformFor` helper to generate
a platform for the (arch, imageForamat, UEFIVendor) tuple.

Note that this also change is not a pure refactor but change
the following behavior:

- `QCOW2Compat: "1.1"` is no longer set because this is the
  default in qemu since 2013 (upstream commit 8ad1898c
  version qemu 1.7)
- uefi vendor is now set for the disk image according to source
  info. This used to be hardcoded for aarch64 and unset for the
  others. This should be mostly irrelevant as only installers
  and the bootc legacy pipeline that uses ostree stages directly
  need it. But having it should do not harm and until PR#689
  is merged we will need for the bootc legacy pipeline.

But if this looks too risky I can revert, adding architectures
happens rarely enough and maybe this refactor is overkill.