bib: resolve container content from the external dnf again
This commit goes back to the approach we had in PR#565 to resolve using
the solver from outside of the container. The reason to back is that
some containers (like f41) do not have the python dnf available so
they break with the "run dnfjson inside" approach (that was meant
as a quick fix only anyway).

Huge kudos to Achilleas for most of the work here.

Co-Authored-By: Achilleas Koutsou <achilleas@koutsou.net>