workflow: move workflow to run on ubuntu-24.04
This commit moves the GH runner to move to 24.04. This means we
no longer need a extra repository to get a current version of
podman and it generally means our software stack is more modern
and closer to what version people run.

Note that this also drops "podman,skopeo" as those are preinstalled.

Unfortuantely the default 24.04 podman configuration is buggy and
a workaround is needed.