progress: add progressbar module
This commit adds a progressbar interface and implementations:
1. terminalProgressBar based on github.com/cheggaaa/pb/v3
2. plainProgressBar to just pass the message but no progress
3. debugProgressBar to print the osbuild/images status

The interface is very tailored to the needs of bib but it should
be similar enough to be reusable by `image-builder-cli`.

It contains a top level spinner that gives the current step:
"Make manifest" or "Building image". The the subprogress from
osbuild: the pipelines and the stages from the pipelines as
progress bars and the last line is the latest status message,
usually something from osbuild like "Starting pipeline qcow2".

The spinenr is a bit of a creative choice, it could also be a
progress bar because we know the steps (manifest+build). But the
runtime is totally dominated by the build step so having a first level
progress bar is odd as the UI will spend 90% of the time in it.
Plus the spinner gives it more a more "dynamic" appearnce.