progress: auto-select progress based on the environment
This commit adds automatic progress bar selection based on the
environment. Ideally we would check for a terminal but with
podman we never have one. So we check for TERM and SSH_CONNECTION
and anf if that is used we assume it's an interactive shell and
give the nice "terminalProgressBar". If that is not set we assuem
we run in a script or CI/CD environment and select plainProgressBar.