github/actions: Enable /jira-epic slash commands
This change allows for using the command to create Jira Tasks under a
given Epic both in a pull request comment or in the pull request
description. By default, this action uses the HMS project to create the
Task in.

Also, the Epic referred to needs to be in a project that is accessible
to the imagebuilder-bot account (currently: COMPOSER, HMS).