test: disable the fedora iso test for now
This commit disables the fedora ISO test for now. The download of
the RPMs tends to be very fragile, often mirrors are outdated or
broken and we get many 404 or 500 errors.

The real fix is of couse to switch to a librepo based downloading
in osbuild but for now we just disable this test so that we can
merge PRs again without having to retry multiple times :(