build-sys: Bump to Fedora 41
Fedora N-1 (40 now) is stable, but will eventually get stale.
In my case we happen to be using this image as a "builder"
image and I added a new feature to rpm-ostree, which I didn't
ship to F40. I may still do so.

However it's clearly a good idea for us to keep updated.

This all said, it's actually not clear to me that Fedora is
the right default base image - it may make sense to target
e.g. c10s or c9s?

This all relates to https://gitlab.com/fedora/bootc/tracker/-/issues/2
as well.

Signed-off-by: Colin Walters <walters@verbum.org>