README.md: update supported flags table
This commit updates the documentation around the supported flags.
It also switches on "bold" for a smaller subset of the important
flags.

Note that still not all flags are shows, some are documented
below (like the `--aws-*`) and some (like --store, --rpmmd)
require more context, i.e. explaining that they require calling
podman with the right `--volume` mappings.