test: add integration test for reporting errors from osbuild
This commit adds a proper integration test that ensures that we
always report the details of a failing osbuild run. This prevents
regressions like the one from #810.

This ensures that:
a) we report the messags from a broken stage
b) we report any osbuild messages as well to catch e.g.
   crashes that are not reported via the json progress

It is archived by creating a new build container fixture that is
deliberately broken, i.e:
a) the org.osbuild.selinux stage is replaced with a fake that
   will echo some msgs and then error
b) osbuild itself is is wrapped around so that we can reliably
   echo some canary strings before the real osbuild is executed