distrodef: fix incorrect detection of `centos-10` vs `10.0`
This commit fixes the issue that when the `os-release` has a
`10.0` in it, a distrodef ending with `-10` will not be identified
as the closest match.

We saw this issue with the `rhel-10.0` beta that has a
`VERSION_ID=10.0` field and a `rhel-10.yaml` distro-def. Here the
code was skipping over rhel-9.yaml and felt back to rhel-9 when
it really should have used `rhel-10.yaml`.

Thanks to Achilleas for tracking this down.