distrodef: fix incorrect detection of `centos-10` vs `10.0`
This commit fixes the issue that when the `os-release` has an
`ID=10.0` in it, a distrodef file ending with `-10` will not
be identified as the closest match because the code is using
a "strictly bigger" relation to sort versions but for semver
`10.0` and `10` are just identical so we need ">=" instead
of ">" (see the actual diff).

We saw this issue with the `rhel-10.0` beta that has a
`VERSION_ID=10.0` field and a `rhel-10.yaml` distro-def. Here the
code was skipping over rhel-10.yaml and felt back to rhel-9 when
it really should have used `rhel-10.yaml`.

Thanks to Achilleas for tracking this down.