workflow: run pytests via matrix to get more parallel runs
This commit moves the test running into a matrix so that
we get more parallel testing. It will still be (much) dominated
by `test_build_iso.py` but at least this way a flaky test in
e.g. `test_container` is much faster to re-run.

With multiple VMs we can probably also parallize the tests
because we have less images per VM to test so diskspace
may be less of an issue.