test: set the cache-ttl to 1h in our tests
The current caching of the test containers can lead
to unexpected results when e.g. the upstream osbuild
repo gets updated but podman caches the relevant line
in the Containerfile because the install of osbuild
has not changed. To counter this the container building
sets an agressive TTL of 1h for the cache.

This fixes the issues locally.