test: boot iso with both bios/uefi (regression test #912)
This commit adds a regression test for issue #912. The
underlying issue was that the images library changed the
API for iso boot options [0] and bib did not get updated.

So we need a test that also boots the ISO with bios to
enusre this can't happen again (we only tested UEFI).

In a followup we should try to move the `manifestForISO`
code into images so that changes inside the images
library are directly done in a single place.

[0] https://github.com/osbuild/images/pull/1289