tools: add gen-manifest helper
Add a `gen-manifests` helper that creates mock manifests from
the given image type. Its not perfect mocks as it still will
look into the bootc ref (centos9).

Its not generating checksums only because to compare from the
manifests generated by bib and images its easier to work with
full manifests and we only have very few of them in bib.