test: remove roots home customizations for ISOs
The root user cannot be customized much with anaconda
and with https://github.com/osbuild/images/pull/1806
we got stricter about it. Unfortunately our own test
manifest sets the root home for ISOs (because we need
it for disks and use the same blueprint) and it
panic()s now.

Fix this by removing the problematic option for ISOs
from the blueprint.