bib: use org.osbild.rhsm for secrets instead of mtls
We currently run depsolve-dnf against the container rootdir.
For subscribed containers this means that we find a bunch
of repositories with ssl keys. For dnfjson this means it
switches the secrets provider to "org.osbuild.mtls" because
it (rightfully) assumes that while the secrets are in this
alternative rootdir they are not generally discoverable on
the system.

However this assumption is mostly untrue in our container world
because for RHEL containers podman will map the host secrets into
the container. So for RHEL containers we can just use the host
secrets and can drop the mTLS setup. This commit does that
and switches the secrets provider from mtls to rhsm (in a
rather hackish way right now).

This is great because the mTLS setup is very disconnected from
the manifest which means on a subscribed machine one cannot
create a rhel-bootc container manifest and then just run
it manually via osbuild. The workaround is to manually create
the mtls env (which is a bit clunky). The same is (probably)
true when moving image-mode into the service. Here manifest
generation and building is decoupled so we would hit the
same issue.

Note that this change is also bad because the existing code
was more general - a bootc container with its own set of
secrets inside the container (but not on the host) will no
longer work with this change applied.

Note that the disconnect between manifest/mtls setup is
also causing issues when moving this code into images.

So far the assumption in images/ibcli is that a manifest is
enough to be able to run a build with osbuild. However for
bootc manifests without something like this change this is
no longer true, we would have to carry the extra mTLS env
around and also add some lifecycle rules to delete the
temporary mTLS dir when its no longer needed.