bib: error for any repo secrets not mapped in from the host
This commit addresses the issue that there could be repository
secrets inside the container that are not mapped from the host.

Such a setup is an error now. Its unclear if there are any
containers out there that uses the (accidential) support for
repository secrets directly from the container to auth against
mtls enabled repos to build ISOs for a bootc image.

If people need this there are two options:
a) put secrets under /etc/pki/entitlement and /etc/rhsm/ca so
   that podman automatically maps it, then things should just
   work as before
b) we could write a new secrets provider org.osbuild.container or
   something that would take a container id and a path and then
   during manifest build we would have to mount the container and
   make the secrets available for curl/librepo this way.

For now lets go with (a). Ulitmately ISO installers should come
as bootc containers and should not be constructed via RPMs. Once
that happens this issue goes away entirely.