deps: switch yaml libraries
During a recent chat it was brought up that `go.pkg.in/yaml.v3` is
unmaintained. The YAML organization forked the unmaintained package and
is now updating it. The `v3` we use here only receives security fixes
while `v4` will eventually support new features and more of YAML.

Let's do the easy an uncontroversial swap first as this is a maintained
drop-in replacement.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>