bib: add support for ID_LIKE when finding the distro YAML
This commit adds a fallback check when no direct match for
the distro ID from the bootc image is found. As long as
the bootc container sets a correct `ID_LIKE` we should
automatically have the same behavior as before when
we used symlinks as aliases.

If needed we might need to reintroduce an explicit
mapping like:
```go
// mapping of distro IDs that are compatible with
var distroCompat = map[string]string{
	"almalinux":     "rhel",
	"aurora":        "fedora",
	"aurora-helium": "rhel",
	"bazzite":       "fedora",
	"bluefin":       "fedora",
	"heliumos":      "rhel",
	"rocky":         "rhel",
}
```
but lets hope we don't need this.

Note that these tests use the "real" distros.yaml from the images library
so make sure that if rhel-9 goes away the test is updated for rhel-XY
(only the version needs to change)