bib: add support for `--type bootc-installer`
This commit adds support for the new `bootc-installer` image
type that will take a bootc container and create an ISO out
of it. It also adds a new `--installer-payload-ref` option
so that the user can specify a different payload container
to install.

See https://github.com/osbuild/images/pull/1906 for details.

This is the equivalent of
https://github.com/osbuild/image-builder-cli/pull/341
for bootc-image-builder and allows us to build these kinds
of images with bib now too.