bib: simplify manifestFromCobraForLegacyISO
This commit merged manifestForISO and manifestFromCobraForLegacyISO
to improve readbility. Usually this is bad but here we have too
many *manifest*() functions and the benefit of having the extra
ManifestConfig struct is not there anymore. So drop all that so
that we only have:
```
manifestFromCobraForLegacyISO()
imgTypeManifestForLegacyISO()
```
left. The later will generate the `manifest.Maniest` only
and is like imgType.Manifest() (but legacyISO is not an imageType
yet).