test: demo flatpak installs via bootc-installer container
Tiny commit that demos a container installer with integrated
flatpaks. Its a bit messy but that is mostly because the
bootc-installer is a bit messy (until we either have a base
container or a script).

It shows we can have installed flatpaks without any new
concepts. Once https://github.com/flatpak/flatpak/pull/6116/
is available this can become even nicer, currently the
install is done via a kickstart with explicit naming of
the flatpaks, when we can use preinstall.d this becomes
fully declarative.

Search for "LOOK HERE:" in the test_build_iso.py for the
relevant parts.